% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaywork.R
\name{jaywork}
\alias{jaywork}
\title{Jay's Routine for Data Analysis}
\usage{
jaywork(file_name)
}
\arguments{
\item{file_name}{File name that ends with .Rmd}
}
\description{
Jay's Routine for Data Analysis
}
\examples{
\dontrun{
  jaywork("analysis.Rmd")
}
}
